Extract mercury response to one file
