--- mercury/model/resp/get_account_info_resp.go
@@ -1,14 +1,0 @@
-package resp
-
-type AccountType string
-
-const (
-	Acp    AccountType = "Acp"
-	PwLock AccountType = "PwLock"
-)
-
-type AccountInfo struct {
-	AccountNumber  uint32      `json:"account_number"`
-	AccountAddress string      `json:"account_address"`
-	AccountType    AccountType `json:"account_type"`
-}
