--- mercury/model/resp/query_transactions_resp.go
@@ -1,13 +1,0 @@
-package resp
-
-type PaginationResponseTransactionView struct {
-	Response   []*TransactionViewWrapper `json:"response"`
-	Count      uint64                    `json:"count,omitempty"`
-	NextCursor uint64                    `json:"next_cursor,omitempty"`
-}
-
-type PaginationResponseTransactionInfo struct {
-	Response   []*TransactionInfoWrapper `json:"response"`
-	Count      uint64                    `json:"count,omitempty"`
-	NextCursor uint64                    `json:"next_cursor,omitempty"`
-}
